테스트 코드 any() 삭제 (#193)
* test : ParticipantGameService
any() 삭제
경기 참가자 리스트 조회 성공 구현

* test : ParticipantGameRepository.countByStatusAndGameId
반환형 int로 변경
그에 따른 test 코드 및 service 수정

* test : 경기 테스트코드 save any() 삭제
해당 Entity에 equals, hashCode 추가

+++ src/main/java/com/zerobase/hoops/gameCreator/repository/ParticipantGameRepository.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Repository;
 public interface ParticipantGameRepository extends
     JpaRepository<ParticipantGameEntity, Long> {
 
-  long countByStatusAndGameId(
+  int countByStatusAndGameId(
       ParticipantGameStatus participantGameStatus, Long gameId);
 
   List<ParticipantGameEntity> findByStatusInAndGameId(
